chore: Separate domain modules
