+++ build.gradle
@@ -15,4 +15,5 @@ subprojects {
     repositories {
         mavenCentral()
     }
+
 }
