+++ core/build.gradle
@@ -27,6 +27,7 @@ repositories {
 
 
 dependencies {
+    implementation project(path: ":domain", configuration: 'default')
     implementation 'org.springframework.boot:spring-boot-starter-data-redis'
     implementation 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'org.springframework.boot:spring-boot-starter-web'
