+++ core/src/main/kotlin/com/baek9/core/register/PushRegisterRepository.kt
@@ -1,6 +1,6 @@
 package com.baek9.core.register
 
-import com.baek9.core.register.entity.ReservedPushRegister
+import com.baek9.domain.register.ReservedPushRegister
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Query
 import java.time.LocalDateTime
