+++ core/src/main/kotlin/com/baek9/core/register/dto/EmailForm.kt
@@ -1,5 +1,6 @@
 package com.baek9.core.register.dto
 
+import com.baek9.domain.register.ReservedPushRegister
 import lombok.Getter
 
 @Getter
@@ -10,9 +11,14 @@ open class EmailForm(
 ) {
     companion object {
         val ERROR: EmailForm = EmailForm("", "", "dltmd202@gmail.com")
+
+        fun from(entity: ReservedPushRegister): EmailForm {
+            return EmailForm(entity.title, entity.message, entity.email)
+        }
     }
 
     override fun toString(): String {
         return "{ \"title\": \"$title\", \"message\": \"$message\", \"email\": \"$email\" }"
     }
+
 }
