+++ core/src/main/kotlin/com/baek9/core/register/dto/ReservedEmailForm.kt
@@ -1,6 +1,6 @@
 package com.baek9.core.register.dto
 
-import com.baek9.core.register.entity.ReservedPushRegister
+import com.baek9.domain.register.ReservedPushRegister
 import lombok.Getter
 import java.time.LocalDateTime
