+++ settings.gradle
@@ -1,2 +1,5 @@
+
 rootProject.name = 'messaing'
 include 'core'
+include 'domain'
+
